% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpie.R
\name{ggpie}
\alias{ggpie}
\title{Draw pie charts}
\usage{
ggpie(
  df,
  .x,
  .y,
  .offset = 1,
  .color = "white",
  .number = TRUE,
  .legend = FALSE,
  .round = 1,
  .percent = TRUE,
  .start = 0,
  .textcolor = "white"
)
}
\arguments{
\item{df}{A data frame.}

\item{.x}{A categorical variable.}

\item{.y}{A numeric variable.}

\item{.offset}{A number, the distance between the center and label.}

\item{.number}{If the number appears in the label.}

\item{.legend}{If the legend is kept.}

\item{.round}{The round number in the label.}

\item{.percent}{If the percent appears in the label.}

\item{.start}{The starting angle.}

\item{.textcolor}{The color of the label.}
}
\description{
Draw pie charts with ggplot2.
}
\examples{
df <- data.frame(type = c("a", "b", "c"), value = c(5, 3, 2))
ggpie(df, type, value)
}
